Set $HOME to build dir /homeless-shelter in non-sandboxed builds
Non-sandboxed builds running as root can easily create
/homeless-shelter. This is an issue since it'll fail the subsequent
nix builds.

Usually, a derivation trying to access $HOME is a code smell. However,
in some cases, like when pip builds a wheel, the software can try to
access $HOME, then fallback to some saner location when it realize it
can't access it. That's an issue for the Nix un-sandboxed builds since
root can indeed access /homeless-shelter.

We set $HOME to $buildDirectory/homeless-shelter to make sure that a
Nix build won't influence the subsequent ones.

I don't think this solution is generic enough to be upstreamed as it
is. We probably rather should try to mimick the sandboxed builds and
set $HOME to a place root cannot access.